refactor(doc): split swagger reverse-proxy into swaggerproxy.go; document K8sSource.GetDiff (review batch 9)
- Move the self-contained reverse-proxy machinery (newProxyHandler, copyHeaders,
  hopByHop, parsedAllowedTarget) out of the 406-line swagger.go into
  swaggerproxy.go. Same package, no behavior/API change.
- Document that K8sSource.GetDiff is intentionally unsupported (k8s observes only
  the deployed revision; the diff endpoint falls back to OCI/local sources).
100% coverage.
